clang-tidy

+++ src/extension/IOController_Posix.hpp
@@ -132,7 +132,8 @@ namespace extension {
                             int bytes_available = 0;
                             bool valid          = ::ioctl(fd.fd, FIONREAD, &bytes_available) == 0;
                             if (valid && bytes_available == 0) {
-                                fd.revents = fd.revents | IO::CLOSE;
+                                // NOLINTNEXTLINE(google-runtime-int)
+                                fd.revents = short(fd.revents | IO::CLOSE);
                             }
                         }
